dns-lease-manager: fix the mingw build and the live lane's resolv.conf
The Windows (MinGW) test build exposed two pre-existing portability bugs in
the fork: lib/fec.cpp cast a pointer to long (32-bit on Windows x64, so the
cast lost precision), now (intptr_t); and xor_spe.S kept its comment outside
the HAVE_PPC_SPE guard, so a non-PowerPC build still fed assembly to the
assembler, which the MinGW toolchain rejects. The guard now covers the whole
file, so it assembles to nothing without SPE.

The live DNS lane writes resolv.conf, but a GitHub runner's /etc/resolv.conf
is managed by systemd-resolved and not writable even as root. The script now
bind-mounts a private resolv.conf over it (root, as the lane requires), torn
down and the original restored in cleanup.

+++ lib/fec.cpp
@@ -45,6 +45,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <stdint.h>
 
 typedef unsigned long u_long;
 /*
@@ -1159,7 +1160,7 @@ fec_free(void *p0)
 {
 	struct fec_parms *p= (struct fec_parms *) p0;
     if (p==NULL ||
-       p->magic != ( ( (FEC_MAGIC ^ p->k) ^ p->n) ^ (int)((long)p->enc_matrix)) ) {
+       p->magic != ( ( (FEC_MAGIC ^ p->k) ^ p->n) ^ (int)((intptr_t)p->enc_matrix)) ) {
 	fprintf(stderr, "bad parameters to fec_free\n");
 	return ;
     }
@@ -1196,7 +1197,7 @@ fec_new(int k, int n)
     retval->dec_matrix = NEW_GF_MATRIX(k, k);
     retval->dec_buf = NULL ;
     retval->dec_buf_sz = 0 ;
-    retval->magic = ( ( FEC_MAGIC ^ k) ^ n) ^ (int)((long)retval->enc_matrix) ;
+    retval->magic = ( ( FEC_MAGIC ^ k) ^ n) ^ (int)((intptr_t)retval->enc_matrix) ;
     tmp_m = NEW_GF_MATRIX(n, k);
     /*
      * fill the matrix with powers of field elements, starting from 0.
